+++ homeassistant/components/tolo/__init__.py
@@ -28,7 +28,7 @@ _LOGGER = logging.getLogger(__name__)
 
 
 async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
-    """Set up tolosauna from a config entry."""
+    """Set up tolo from a config entry."""
     coordinator = ToloSaunaUpdateCoordinator(hass, entry)
     await coordinator.async_config_entry_first_refresh()
