+++ homeassistant/components/tolo/strings.json
