+++ homeassistant/components/tolo/translations/en.json
