+++ mypy.ini
@@ -1441,7 +1441,7 @@ no_implicit_optional = true
 warn_return_any = true
 warn_unreachable = true
 
-[mypy-homeassistant.components.tolosauna.*]
+[mypy-homeassistant.components.tolo.*]
 check_untyped_defs = true
 disallow_incomplete_defs = true
 disallow_subclassing_any = true
