+++ requirements_all.txt
@@ -2313,7 +2313,7 @@ tmb==0.0.4
 # homeassistant.components.todoist
 todoist-python==8.0.0
 
-# homeassistant.components.tolosauna
+# homeassistant.components.tolo
 tololib==0.1.0b2
 
 # homeassistant.components.toon
